Add gang_related to call_general_category view
Added additional regexes to identify gang-related calls based on
their notes.  Note that these don't correlate especially well
with the DPD's existing "gang_related" column in the incident
table; many of the DPD-identified gang-related calls don't have
anything that would indicate gang activity/affiliation in their
notes.  Most of the regex-identified calls are calls on which
a residence turns up a "suspected gang member" or "validated gang
member" alert for one of its occupants.  I'm not sure how useful
this category will be for the DPD because of that.