Perform sentiment analysis on call notes
Performed sentiment analysis using both the ANEW and AFINN dictionaries
on the call notes, attempting to find a way to flag calls that indicate
negative interactions between officers and citizens.  This was
unsuccessful, as the sentiment scores tend to be too general and too
sensitive to domain words/phrases such as "suicide" or "road rage" to
aid in this kind of classification.  The better approach will be to
use a regex tuned to specific words and phrases.