Add ofcr/ctzn conflict to general call categories
We can use a regex to parse the call notes and identify calls
where some kind of conflict or negative interaction has occurred
between a police officer and a citizen ("officer-citizen conflict",
or OCC).

We do this by searching for either a phrase meaning a
citizen wishes to file a complaint against an officer or both a
phrase identifying an officer and a phrase indicating some kind
of conflict (such as "rude", "upset", or "disrespect").  It's
impossible to estimate the recall of this approach, as there are
many different ways to describe an OCC in the call notes, but I
would estimate that the precision is high enough to be useful
(80-90%).  This flag could be best used to allow supervisors
to keep an eye on any calls that negatively impact their
relationship with the community (and possibly identify problem
officers if there are repeat offenders), as opposed to providing
a means of determining exactly how many OCC calls occur.