Add response_time to call
This field was derived via (first_unit_arrive - time_received).
It will be null if either the first_unit_arrive timestamp is
null (~105k records) or the time_received timestamp is null
(should never happen and isn't in our data so far).

Note that sometimes the response_time is negative because
the calltakers occasionally screw up recording the timestamps
of the various events.  This is rare, though (28 out of 200k
non-null response times).