Add homeless to general categories
Using most of the same keywords from Tampa, I added 'homeless'
to the group of general categories.  Due to the inclusion of
terms like 'begging' and 'panhandling', this really represents
more of a combination of 'homeless' and 'urban poor'.

I decided not to include the keyword 'shelter', which was used
in Tampa, because it caused too many false positives from
phrases like 'women's shelter', 'domestic violence shelter',
'animal shelter', and others.