Converted filtering to use django-url-filter
django-url-filter:
https://django-url-filter.readthedocs.org/en/latest/index.html

This allows us much greater flexibility in what we can filter. We can now
filter for the non-equality of something as well as inclusion in a group.

What we lose is the ability to filter based on a function. We can add this back
by extending the library in the future.