First version of OfficerActivity model/filter
This model transforms the call_log data in such a way that
we can display officer allocation information based on the
in_call view in the database.  The filter is fairly well
tested and appears to be working, but the model tests are
currently failing due to some difficulty filtering based on
calls, and some more need to be added.