Increase postgres work_mem
For the aggregation required by the OfficerActivity view,
we have to raise work_mem above ~12MB (set at 16 to be safe) in order
to avoid performing a sort on disk.  Without setting the work_mem
above this level, the query takes about 20 secs; with the proper
setting, it takes around 1.5.